[CSS] Apply first line style even if there are out-of-flow blocks
This patch means the code change to apply first line style on the first
innermost in-flow block even if there are out-of-flow blocks.
Following the spec (https://drafts.csswg.org/css-pseudo-4/#first-line-pseudo),
The first formatted line of an element must occur inside a block-level
descendant in the same flow (that is not out-of-flow due to floating or
positioning). So, we should skip out-of-flow children in
LayoutBlock::EnclosingFirstLineStyleBlock().

Bug: 979253
Change-Id: I58836b223b58816f731096a7a9f93e612278d42f