Test promise structure and options changing in Pointer Lock
Adding tentative Web Platform Tests for the new Pointer Lock features.

The previous test framework did not check for user gesture requirements.
This tests adds changes to the Web Test Shell to mimic the
implementation. Plans to move the logic in the implementation so that it
is used in the tests have been made here: https://crbug.com/1140673.

Bug:982379
Change-Id: I798315cf14d76ee1b9818eafd89a9c6d3b8613b6