HTMLPluginElement: Add object to frame view's update part if locked.
This patch ensures that when we delay the load because the plugins
need to be loaded and we're under a content-visibility lock, then
we add the object to be updated into frame view's list immediately

This is typically done during layout, but if layout is prevented then
we run into a possibility of delaying load event forever. Adding
the object to the update set immediately avoids the problem.

R=masonfreed@chromium.org

Bug: 1138128
Change-Id: I4d918c6bbdcd80aa6f01bf3b068acb64a93b0196